+++ src/main/java/com/teamtea/eclipticseasons/common/item/info/GrowthInfo.java
@@ -1,34 +1,47 @@
 package com.teamtea.eclipticseasons.common.item.info;
 
-import com.teamtea.eclipticseasons.api.constant.biome.Humidity;
-import com.teamtea.eclipticseasons.api.constant.solar.Season;
-import com.teamtea.eclipticseasons.api.data.climate.AgroClimaticZone;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Holder;
+import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.level.block.state.BlockState;
-import org.jspecify.annotations.Nullable;
-
-import java.util.List;
+import net.minecraft.network.chat.ComponentSerialization;
+import net.minecraft.network.codec.ByteBufCodecs;
+import net.minecraft.network.codec.StreamCodec;
 
 public record GrowthInfo(
         BlockPos pos,
-        BlockState state,
+        // BlockState state,
         Component cropName,
 
-        @Nullable Holder<AgroClimaticZone> agroClimaticZone,
+        // @Nullable Holder<AgroClimaticZone> agroClimaticZone,
 
-        boolean greenhouse,
+        // boolean greenhouse,
         int greenhouseLevel,
 
         float growChance,
-        int growChanceLevel,
+        // int growChanceLevel,
 
         boolean needsSeasonCore,
-        boolean humidityMismatch,
+        boolean humidityMismatch
 
-        float humidity,
-        @Nullable List<Season> likedSeasons,
-        @Nullable List<Humidity> likedHumidity
+        // float humidity,
+        // @Nullable List<Season> likedSeasons,
+        // @Nullable List<Humidity> likedHumidity
 ) {
+
+
+    public static final StreamCodec<RegistryFriendlyByteBuf, GrowthInfo> STREAM_CODEC = StreamCodec.composite(
+            BlockPos.STREAM_CODEC,
+            GrowthInfo::pos,
+            ComponentSerialization.STREAM_CODEC,
+            GrowthInfo::cropName,
+            ByteBufCodecs.VAR_INT,
+            GrowthInfo::greenhouseLevel,
+            ByteBufCodecs.FLOAT,
+            GrowthInfo::growChance,
+            ByteBufCodecs.BOOL,
+            GrowthInfo::needsSeasonCore,
+            ByteBufCodecs.BOOL,
+            GrowthInfo::humidityMismatch,
+            GrowthInfo::new
+    );
 }