+++ src/main/java/com/teamtea/eclipticseasons/common/network/SimpleNetworkHandler.java
@@ -1,5 +1,8 @@
 package com.teamtea.eclipticseasons.common.network;
 
+import com.teamtea.eclipticseasons.common.core.map.MapChecker;
+import com.teamtea.eclipticseasons.common.item.info.GrowthInfoResolver;
+import com.teamtea.eclipticseasons.common.network.clientmesage.GrowthInfoQuery;
 import com.teamtea.eclipticseasons.common.network.message.*;
 import com.teamtea.eclipticseasons.config.sync.ESConfigFilePayload;
 import com.teamtea.eclipticseasons.config.sync.ESConfigSync;
@@ -13,6 +16,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 
 import java.util.Collection;
+import java.util.Optional;
 
 public final class SimpleNetworkHandler {
 
@@ -25,6 +29,9 @@ public final class SimpleNetworkHandler {
         PayloadTypeRegistry.clientboundPlay().register(ChunkBiomeUpdateMessage.TYPE, ChunkBiomeUpdateMessage.STREAM_CODEC);
         PayloadTypeRegistry.clientboundPlay().register(HumidModifyMessage.TYPE, HumidModifyMessage.STREAM_CODEC);
         PayloadTypeRegistry.clientboundPlay().register(UpdateTempChangeMessage.TYPE, UpdateTempChangeMessage.STREAM_CODEC);
+        PayloadTypeRegistry.clientboundPlay().register(GrowthInfoMessage.TYPE, GrowthInfoMessage.STREAM_CODEC);
+        PayloadTypeRegistry.serverboundPlay().register(GrowthInfoQuery.TYPE, GrowthInfoQuery.STREAM_CODEC);
+
 
         PayloadTypeRegistry.clientboundConfiguration().register(ESConfigFilePayload.TYPE, ESConfigFilePayload.STREAM_CODEC);
         PayloadTypeRegistry.clientboundPlay().register(ESConfigToClientPayload.TYPE, ESConfigToClientPayload.STREAM_CODEC);
@@ -38,6 +45,18 @@ public final class SimpleNetworkHandler {
                         if (player instanceof ServerPlayer serverPlayer)
                             ESConfigSync.INSTANCE.syncToSever(payload, serverPlayer);
                     });
+                }
+        );
+
+        ServerPlayNetworking.registerGlobalReceiver(
+                GrowthInfoQuery.TYPE,
+                (payload, context) -> {
+                    Player player = context.player();
+                    if (player instanceof ServerPlayer serverPlayer
+                            && MapChecker.isLoadedOnlyServer(serverPlayer.level(), payload.getPos()))
+                        send(serverPlayer, new GrowthInfoMessage(Optional.ofNullable(GrowthInfoResolver.resolve(
+                                serverPlayer.level(), payload.getPos(), serverPlayer.level().getBlockState(payload.getPos())
+                        ))));
                 }
         );
     }